--- src/Components/Comment2.jsx/Action.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Action = ({ handleClick, type, className }) => {
-  return (
-    <div className={className} onClick={handleClick}>
-      {type}
-    </div>
-  );
-};
-
-export default Action;
