--- src/Components/Comment2.jsx/CommentApp.jsx
@@ -1,42 +1,0 @@
-import { useState } from "react";
-import Comment from "./Comment";
-import useNode from "../Hook/useNode";
-
-const comments = {
-  id: 1,
-  items: [],
-};
-const CommentApp = () => {
-  const [commentsData, setCommentsData] = useState(comments);
-
-  const { insertNode, editNode, deleteNode } = useNode();
-
-  const handleInsertNode = (folderId, item) => {
-    const finalStructure = insertNode(commentsData, folderId, item);
-    setCommentsData(finalStructure);
-  };
-
-  const handleEditNode = (folderId, value) => {
-    const finalStructure = editNode(commentsData, folderId, value);
-    setCommentsData(finalStructure);
-  };
-
-  const handleDeleteNode = (folderId) => {
-    const finalStructure = deleteNode(commentsData, folderId);
-    const temp = { ...finalStructure };
-    setCommentsData(temp);
-  };
-
-  return (
-    <div className="App">
-      <Comment
-        handleInsertNode={handleInsertNode}
-        handleEditNode={handleEditNode}
-        handleDeleteNode={handleDeleteNode}
-        comment={commentsData}
-      />
-    </div>
-  );
-};
-
-export default CommentApp;
