--- src/GarageComment2.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import CommentApp from './Components/Comment2.jsx/CommentApp'
-
-export default function Garage() {
-  return (
-    <div>
-        <CommentApp />
-    </div>
-  )
-}
